requirements.txt
# LiveKit Protocol Agent - Requirements
# Install with: pip install -r requirements.txt

# Core LiveKit Agents Framework
livekit-agents>=1.3.0

# Speech-to-Text Provider (Deepgram - recommended)
livekit-plugins-deepgram>=1.3.0

# Voice Activity Detection (Silero)
livekit-plugins-silero>=1.3.0

# Environment Variables
python-dotenv>=1.0.0

# Async File I/O
aiofiles>=23.0.0

# Optional: Alternative STT Providers (uncomment to use)
# livekit-plugins-speechmatics>=1.3.0
# livekit-plugins-openai>=1.3.0
